product admin

+++ products/views.py
@@ -2,18 +2,17 @@ from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
-from django.db.models.functions import Lower 
+from django.db.models.functions import Lower
+
 from .models import Product, Category
 from .forms import ProductForm
 
 # Create your views here.
 
-@login_required
+
 def all_products(request):
-    """ A view to show all products, including sorting and search queries """
-    if not request.user.is_superuser: 
-        messages.error(request, 'Sorry, only store owner can do that')
-        return redirect(reverse('home'))
+    """A view to show all products, including sorting and search queries"""
+
     products = Product.objects.all()
     query = None
     categories = None
@@ -27,12 +26,13 @@ def all_products(request):
             if sortkey == 'name':
                 sortkey = 'lower_name'
                 products = products.annotate(lower_name=Lower('name'))
-
+            if sortkey == 'category':
+                sortkey = 'category__name'
             if 'direction' in request.GET:
                 direction = request.GET['direction']
                 if direction == 'desc':
                     sortkey = f'-{sortkey}'
-                products = products.order_by(sortkey)
+            products = products.order_by(sortkey)
 
         if 'category' in request.GET:
             categories = request.GET['category'].split(',')
@@ -42,10 +42,14 @@ def all_products(request):
         if 'q' in request.GET:
             query = request.GET['q']
             if not query:
-                messages.error(request, "You didn't enter any search criteria!")
+                messages.error(
+                    request, "You didn't enter any search criteria!"
+                )
                 return redirect(reverse('products'))
 
-            queries = Q(name__icontains=query) | Q(description__icontains=query)
+            queries = Q(name__icontains=query) | Q(
+                description__icontains=query
+            )
             products = products.filter(queries)
 
     current_sorting = f'{sort}_{direction}'
@@ -60,7 +64,6 @@ def all_products(request):
     return render(request, 'products/products.html', context)
 
 
-@login_required
 def product_detail(request, product_id):
     """A view to show individual product details"""
 
@@ -75,18 +78,25 @@ def product_detail(request, product_id):
 
 @login_required
 def add_product(request):
-    """Add product to the store"""
+    """Add a product to the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
             product = form.save()
-            messages.success(request, 'Succesfully added product')
+            messages.success(request, 'Successfully added product!')
             return redirect(reverse('product_detail', args=[product.id]))
         else:
-            messages.error(request, 'Failed to add product')
+            messages.error(
+                request,
+                'Failed to add product. Please ensure the form is valid.',
+            )
     else:
-
         form = ProductForm()
+
     template = 'products/add_product.html'
     context = {
         'form': form,
@@ -98,6 +108,10 @@ def add_product(request):
 @login_required
 def edit_product(request, product_id):
     """Edit a product in the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -125,8 +139,12 @@ def edit_product(request, product_id):
 
 @login_required
 def delete_product(request, product_id):
-    """Delete a product in the store"""
+    """Delete a product from the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
-    messages.success(request, 'Successfully deleted product!')
+    messages.success(request, 'Product deleted!')
     return redirect(reverse('products'))
